+++ src/main/java/org/reactivesource/AbstractConnectionProvider.java
@@ -5,6 +5,8 @@
  ******************************************************************************/
 package org.reactivesource;
 
+import org.reactivesource.exceptions.DataAccessException;
+
 import java.sql.Connection;
 import java.sql.DriverManager;
 import java.sql.SQLException;
