+++ src/main/java/org/reactivesource/exceptions/ConfigurationException.java
@@ -1,4 +1,4 @@
-package org.reactivesource;
+package org.reactivesource.exceptions;
 
 public class ConfigurationException extends RuntimeException {
     private static final long serialVersionUID = -7104365629425825102L;
