+++ src/main/java/org/reactivesource/psql/PsqlEventSource.java
@@ -19,7 +19,7 @@ import java.util.Map;
 import org.postgresql.PGConnection;
 import org.postgresql.PGNotification;
 import org.reactivesource.ConnectionProvider;
-import org.reactivesource.DataAccessException;
+import org.reactivesource.exceptions.DataAccessException;
 import org.reactivesource.Event;
 import org.reactivesource.EventSource;
 import org.reactivesource.common.JdbcUtils;
