+++ src/test/java/org/reactivesource/psql/PsqlConfiguratorIntegrationTest.java
@@ -22,9 +22,9 @@ import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.sql.Statement;
 
-import org.reactivesource.ConfigurationException;
+import org.reactivesource.exceptions.ConfigurationException;
 import org.reactivesource.ConnectionProvider;
-import org.reactivesource.DataAccessException;
+import org.reactivesource.exceptions.DataAccessException;
 import org.testng.annotations.AfterMethod;
 import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
